settings and updates to db table

+++ db/local.sql
@@ -888,7 +888,7 @@ CREATE TABLE `transaction` (
   `comments` varchar(500) NOT NULL,
   `action` char(50) NOT NULL,
   `type` char(20) NOT NULL DEFAULT '',
-  `value` bigint(50) NOT NULL,
+  `value` decimal(20,2) NOT NULL,
   `dateCreated` bigint(50) NOT NULL,
   `dateUpdated` bigint(50) NOT NULL,
   `order_id` varchar(50) NOT NULL,
@@ -902,7 +902,7 @@ CREATE TABLE `transaction` (
 
 LOCK TABLES `transaction` WRITE;
 /*!40000 ALTER TABLE `transaction` DISABLE KEYS */;
-INSERT INTO `transaction` VALUES (1,'tKoPYt3m9uKzYsiztrdGQ6dQ69E1on1518409422524','','','CONFIRM_PAYMENT',1200,1518409422524,1518409422524,'1');
+INSERT INTO `transaction` VALUES (1,'tKoPYt3m9uKzYsiztrdGQ6dQ69E1on1518409422524','','','CONFIRM_PAYMENT',1200.00,1518409422524,1518409422524,'1');
 /*!40000 ALTER TABLE `transaction` ENABLE KEYS */;
 UNLOCK TABLES;
